Rename from FluxEntity to EntityFlux
Fix the apply to use Kustomize.
