+++ packages/app/src/App.tsx
@@ -35,7 +35,7 @@ import { AppRouter, FlatRoutes } from '@backstage/core-app-api';
 import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
 import { RequirePermission } from '@backstage/plugin-permission-react';
 import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
-import { FluxEntityHelmReleasesCard } from '@weaveworksoss/backstage-plugin-flux';
+import { EntityFluxHelmReleasesCard } from '@weaveworksoss/backstage-plugin-flux';
 
 const app = createApp({
   components: {
@@ -110,7 +110,7 @@ const routes = (
     </Route>
     <Route path="/settings" element={<UserSettingsPage />} />
     <Route path="/catalog-graph" element={<CatalogGraphPage />} />
-    <Route path="/weaveworks-flux" element={<FluxEntityHelmReleasesCard />} />
+    <Route path="/weaveworks-flux" element={<EntityFluxHelmReleasesCard />} />
   </FlatRoutes>
 );
