+++ plugins/backstage-plugin-flux/README.md
@@ -2,7 +2,7 @@
 
 The Flux plugin for Backstage provides views of [Flux](https://fluxcd.io/) resources available in Kubernetes clusters.
 
-![FluxEntitySourcesCard](sources_card.png)
+![EntityFluxSourcesCard](sources_card.png)
 
 ## Content
 
@@ -12,16 +12,16 @@ All cards use the Backstage Kubernetes ["common label"](https://backstage.io/doc
 
 These Cards provide unified views of their resources.
 
-- FluxEntityDeploymentsCard - Displays associated Kustomizations and HelmReleases
-- FluxEntitySourcesCard - Displays associated GitRepositories, OCIRepositories and HelmRepositories
+- EntityFluxDeploymentsCard - Displays associated Kustomizations and HelmReleases
+- EntityFluxSourcesCard - Displays associated GitRepositories, OCIRepositories and HelmRepositories
 
 You can also add cards for resources with the following components, each of these shows specific resources associated to the Entity.
 
-- FluxEntityHelmReleasesCard
-- FluxEntityKustomizationsCard
-- FluxEntityGitRepositoriesCard
-- FluxEntityOCIRepositoriesCard
-- FluxEntityHelmRepositoriesCard
+- EntityFluxHelmReleasesCard
+- EntityFluxKustomizationsCard
+- EntityFluxGitRepositoriesCard
+- EntityFluxOCIRepositoriesCard
+- EntityFluxHelmRepositoriesCard
 
 As with other Backstage plugins, you can compose the UI you need.
 
@@ -123,7 +123,7 @@ The Flux plugins provide several different Cards, which are composable into your
 
 ```tsx
 // In packages/app/src/components/catalog/EntityPage.tsx
-import { FluxEntityHelmReleasesCard } from '@weaveworksoss/backstage-plugin-flux';
+import { EntityFluxHelmReleasesCard } from '@weaveworksoss/backstage-plugin-flux';
 
 // You can add the tab to any number of pages, the service page is shown as an
 // example here
@@ -131,7 +131,7 @@ const overviewContent = (
   <Grid container spacing={3} alignItems="stretch">
     // ...
     <Grid item md={4} xs={12}>
-      <FluxEntityHelmReleasesCard />
+      <EntityFluxHelmReleasesCard />
     </Grid>
     // ...
   </Grid>
@@ -142,7 +142,7 @@ The card has a `many` property which is `true` by default. If your card table is
 
 ```tsx
 <Grid item md={4} xs={12}>
-  <FluxEntityHelmReleasesCard many={false} />
+  <EntityFluxHelmReleasesCard many={false} />
 </Grid>
 ```
 
@@ -151,10 +151,10 @@ The card has a `many` property which is `true` by default. If your card table is
 ```tsx
 // In packages/app/src/components/catalog/EntityPage.tsx
 import {
-  FluxEntityHelmReleasesCard,
-  FluxEntityGitRepositoriesCard,
-  FluxEntityOCIRepositoriesCard,
-  FluxEntityHelmRepositoriesCard,
+  EntityFluxHelmReleasesCard,
+  EntityFluxGitRepositoriesCard,
+  EntityFluxOCIRepositoriesCard,
+  EntityFluxHelmRepositoriesCard,
 } from '@weaveworksoss/backstage-plugin-flux';
 
 const serviceEntityPage = (
@@ -162,16 +162,16 @@ const serviceEntityPage = (
     // ...
     <Grid container spacing={3} alignItems="stretch">
       <Grid item md={12}>
-        <FluxEntityHelmReleasesCard />
+        <EntityFluxHelmReleasesCard />
       </Grid>
       <Grid item md={12}>
-        <FluxEntityHelmRepositoriesCard />
+        <EntityFluxHelmRepositoriesCard />
       </Grid>
       <Grid item md={12}>
-        <FluxEntityGitRepositoriesCard />
+        <EntityFluxGitRepositoriesCard />
       </Grid>
       <Grid item md={12}>
-        <FluxEntityOCIRepositoriesCard />
+        <EntityFluxOCIRepositoriesCard />
       </Grid>
     </Grid>
     // ...
