+++ plugins/backstage-plugin-flux/src/components/EntityFluxDeploymentsCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxDeploymentsCard } from './EntityFluxDeploymentsCard';
