+++ plugins/backstage-plugin-flux/src/components/EntityFluxGitRepositoriesCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxGitRepositoriesCard } from './EntityFluxGitRepositoriesCard';
