+++ plugins/backstage-plugin-flux/src/components/EntityFluxHelmReleasesCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxHelmReleasesCard } from './EntityFluxHelmReleasesCard';
