+++ plugins/backstage-plugin-flux/src/components/EntityFluxHelmRepositoriesCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxHelmRepositoriesCard } from './EntityFluxHelmRepositoriesCard';
