+++ plugins/backstage-plugin-flux/src/components/EntityFluxKustomizationsCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxKustomizationsCard } from './EntityFluxKustomizationsCard';
