+++ plugins/backstage-plugin-flux/src/components/EntityFluxOCIRepositoriesCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxOCIRepositoriesCard } from './EntityFluxOCIRepositoriesCard';
