+++ plugins/backstage-plugin-flux/src/components/EntityFluxSourcesCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityFluxSourcesCard } from './EntityFluxSourcesCard';
