--- plugins/backstage-plugin-flux/src/components/FluxEntityDeploymentsCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntityDeploymentsCard } from './FluxEntityDeploymentsCard';
