--- plugins/backstage-plugin-flux/src/components/FluxEntityGitRepositoriesCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntityGitRepositoriesCard } from './FluxEntityGitRepositoriesCard';
