--- plugins/backstage-plugin-flux/src/components/FluxEntityHelmReleasesCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntityHelmReleasesCard } from './FluxEntityHelmReleasesCard';
