--- plugins/backstage-plugin-flux/src/components/FluxEntityHelmRepositoriesCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntityHelmRepositoriesCard } from './FluxEntityHelmRepositoriesCard';
