--- plugins/backstage-plugin-flux/src/components/FluxEntityKustomizationsCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntityKustomizationsCard } from './FluxEntityKustomizationsCard';
