--- plugins/backstage-plugin-flux/src/components/FluxEntityOCIRepositoriesCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntityOCIRepositoriesCard } from './FluxEntityOCIRepositoriesCard';
