--- plugins/backstage-plugin-flux/src/components/FluxEntitySourcesCard/index.ts
@@ -1,1 +1,0 @@
-export { FluxEntitySourcesCard } from './FluxEntitySourcesCard';
