+++ plugins/backstage-plugin-flux/src/components/index.ts
@@ -1,5 +1,5 @@
-export * from './FluxEntityGitRepositoriesCard';
-export * from './FluxEntityHelmReleasesCard';
-export * from './FluxEntityOCIRepositoriesCard';
-export * from './FluxEntityKustomizationsCard';
-export * from './FluxEntityHelmRepositoriesCard';
+export * from './EntityFluxGitRepositoriesCard';
+export * from './EntityFluxHelmReleasesCard';
+export * from './EntityFluxOCIRepositoriesCard';
+export * from './EntityFluxKustomizationsCard';
+export * from './EntityFluxHelmRepositoriesCard';
